RAG-Based-Application/Url-Data/requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchainhub
unstructured
nltk